Refactoring StopRatioDonut
