+++ traffic_stops/static/js/app/states/nc/Stops.js
@@ -34,72 +34,20 @@ export var StopsHandler = DataHandlerBase.extend({
   }
 });
 
-export var StopRatioDonut = VisualBase.extend({
+export var StopRatioDonut = C.StopRatioDonutBase.extend({
   defaults: {
     showEthnicity: false,
     width: 300,
     height: 300
   },
-  setDefaultChart: function(){
-    this.chart = nv.models.pieChart()
-      .x(function(d){ return d.key; })
-      .y(function(d){ return d.value; })
-      .color(function(d){ return d.data.color; })
-      .width(this.get("width"))
-      .height(this.get("height"))
-      .showLabels(true)
-      .labelType("percent")
-      .donutRatio(0.35)
-      .labelThreshold(0.05)
-      .donut(true)
-      .tooltipContent((key, y, e, graph) => (
-        `<h3 class="stops donut-label">${ key }</h3><p>${ y.replace(/\.\d*/, '') }</p>`
-      ));
-  },
-  drawStartup: function(){
 
-    // get year options for pulldown menu
-    var selector = $('<select>'),
-        year_options = this.data.pie.keys(),
-        opts = year_options.map((v) => `<option value="${v}">${v}</option>`),
-        getData = () => {
-          var value = selector.val();
-          this.dataset =  this.data.pie.get(value);
-          this.drawChart();
-        };
-
-    selector
-      .append(opts)
-      .val("Total")
-      .on('change', getData);
-
-    $('<div>')
-      .html(selector)
-      .appendTo(this.div);
-
-    getData();
-  },
-  drawChart: function(){
-    var data = this._formatData();
-
-    nv.addGraph(() => {
-      d3.select(this.svg[0])
-          .datum(data)
-        .transition().duration(1200)
-          .attr('width', "100%")
-          .attr('height', "100%")
-          .attr("preserveAspectRatio", "xMinYMin")
-          .attr('viewBox', `0 0 ${this.get('width')} ${this.get('height')}`)
-          .call(this.chart);
-    });
-  },
-  _formatData: function(){
+  _formatData: function () {
     var data = [],
         selected = this.dataset,
         items = (this.get('showEthnicity')) ? Stops.ethnicities : Stops.races;
 
     // build data specifically for this pie chart
-    items.forEach(function(d, i){
+    items.forEach((d, i) => {
       data.push({
         "key": Stops.pprint.get(d),
         "value": selected.get(d),
@@ -109,7 +57,8 @@ export var StopRatioDonut = VisualBase.extend({
 
     return data;
   },
-  triggerRaceToggle: function(e, v){
+
+  triggerRaceToggle: function (e, v) {
     this.set('showEthnicity', v);
     this.drawChart();
   }
